refactor(events): move standalone task modules into the events/tasks package
Three events Celery-task modules sat beside the new events/tasks/ package
(workarounds from when events/tasks.py was at the 1000-line gate and they needed
manual registration via EventsConfig.ready):
- events/tasks_stripe.py        -> events/tasks/stripe_webhooks.py
- events/announcement_tasks.py  -> events/tasks/announcements.py
- events/revenue_tasks.py       -> events/tasks/revenue.py

Now that events/tasks is a package, these belong inside it and register through
Celery autodiscovery like every other submodule — so the manual import in
EventsConfig.ready is removed (it imported only these three plus signals; signals
stays). Verified at runtime: all five tasks register via autodiscovery with their
original names.

Pure move, no behavior change:
- All tasks keep their explicit names (events.prune_stripe_webhook_events,
  events.send_scheduled_announcements, events.resend_announcements_to_new_signups,
  events.generate_revenue_report, events.send_scheduled_revenue_reports), so the
  Celery-beat schedules in migrations 0079/0082/0085 (which reference tasks by name
  string) are unaffected.
- events/tasks/__init__.py re-exports the five tasks; the three external import
  sites now use the package path (the announcements relative import .models ->
  events.models). Stale docstrings referencing the old location/ready() wiring updated.

Migration safety swept: no migration imports any task module by Python path or via
import_string/importlib; every migration->task link is a preserved name string.
Full 28-name beat sweep resolves; make check green (ruff, mypy --strict, migrations,
i18n, file-length); 51 affected task/service tests pass.
